# Stack<T> optimization of (Try)Peek, (Try)Pop and Push