# Description

Enabled RCE on array-access and avoided the explicit check for `_size == 0` --> this is done implicitly in the "RCE-if". So some `cmp`s can be saved.

By _Pop_ the effect on value types is not so big, than for reference types (two array accesses).

This PR is a kind of extension to https://github.com/dotnet/corefx/issues/17318

# Benchmarks

## Notes

Code for benchmarks lives [here](https://github.com/gfoidl/Benchmarks/tree/master/corefx/System/Collections/Generic/Stack/source/StackBenchmarks)

Due the use of http://benchmarkdotnet.org the benchmarks were done a couple of times, because some crazy results with perf x2 were reported and this seems too strange. The results shown here are the more realistic ones. Individual results are in the linked repo above.  
The changes from this PR never showed a decrease in perf.

## Peek

``` ini

BenchmarkDotNet=v0.10.11, OS=ubuntu 16.04
Processor=Intel Xeon CPU 2.60GHz, ProcessorCount=2
.NET Core SDK=2.1.3
  [Host]     : .NET Core 2.0.4 (Framework 4.6.0.0), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.4 (Framework 4.6.0.0), 64bit RyuJIT


```
|       Method |     Mean |     Error |    StdDev | Scaled | ScaledSD |
|------------- |---------:|----------:|----------:|-------:|---------:|
| Peek_Default | 1.974 ns | 0.1229 ns | 0.2020 ns |   1.00 |     0.00 |
|     Peek_New | 1.600 ns | 0.1231 ns | 0.2023 ns |   0.82 |     0.13 |


## TryPeek

``` ini

BenchmarkDotNet=v0.10.11, OS=ubuntu 16.04
Processor=Intel Xeon CPU 2.60GHz, ProcessorCount=2
.NET Core SDK=2.1.3
  [Host]     : .NET Core 2.0.4 (Framework 4.6.0.0), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.4 (Framework 4.6.0.0), 64bit RyuJIT


```
|          Method |     Mean |     Error |    StdDev |   Median | Scaled | ScaledSD |
|---------------- |---------:|----------:|----------:|---------:|-------:|---------:|
| TryPeek_Default | 5.212 ns | 0.1905 ns | 0.3977 ns | 5.178 ns |   1.00 |     0.00 |
|     TryPeek_New | 3.982 ns | 0.1716 ns | 0.4640 ns | 3.798 ns |   0.77 |     0.11 |


## Pop

``` ini

BenchmarkDotNet=v0.10.11, OS=ubuntu 16.04
Processor=Intel Xeon CPU 2.60GHz, ProcessorCount=2
.NET Core SDK=2.1.3
  [Host]     : .NET Core 2.0.4 (Framework 4.6.0.0), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.4 (Framework 4.6.0.0), 64bit RyuJIT


```
|      Method |      Mean |     Error |    StdDev | Scaled | ScaledSD |
|------------ |----------:|----------:|----------:|-------:|---------:|
| Pop_Default | 1.0609 ns | 0.0673 ns | 0.0630 ns |   1.00 |     0.00 |
|     Pop_New | 0.8681 ns | 0.1036 ns | 0.0918 ns |   0.82 |     0.10 |


## TryPop

``` ini

BenchmarkDotNet=v0.10.11, OS=ubuntu 16.04
Processor=Intel Xeon CPU 2.60GHz, ProcessorCount=2
.NET Core SDK=2.1.3
  [Host]     : .NET Core 2.0.4 (Framework 4.6.0.0), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.4 (Framework 4.6.0.0), 64bit RyuJIT


```
|         Method |     Mean |     Error |    StdDev | Scaled | ScaledSD |
|--------------- |---------:|----------:|----------:|-------:|---------:|
| TryPop_Default | 3.180 ns | 0.0644 ns | 0.0571 ns |   1.00 |     0.00 |
|     TryPop_New | 2.736 ns | 0.1200 ns | 0.1123 ns |   0.86 |     0.04 |


# Notes

## Push

I did some trials for `Push` too, but the results weren't satisfying. Sometimes it was faster, sometimes slower. On Linux it was nearly always slower. So I reverted the change. 
Changes were done in the commit https://github.com/gfoidl/corefx/commit/012333094dd2b2052eaf6daebaaba2a98f25d2b1

## Inlining

`MethodImplOptions.AggressiveInlining` would give a perf win on the benchmarks, because they are micro-benchmarks. Due to inlining the callsite gets bigger and in real world scenarios the perf might decrease. I ran a benchmark where this happended, unfortunately I can't show the code because it's proprietary.

From https://github.com/dotnet/corefx/pull/12094#issuecomment-249834575

> But we shouldn't add AggressiveInlining unless there's a strong performance-motivated scenario for doing so. Getting something to inline isn't a goal in and of itself.

On the other side `List.Add` has AggressiveInlining -- cf. https://github.com/dotnet/coreclr/blob/master/src/mscorlib/shared/System/Collections/Generic/List.cs#L225

I would stick to not aggressive inline, but what to do?