@gfoidl 2017-12-28
[A Test fails](https://mc.dot.net/#/user/gfoidl/pr~2Fjenkins~2Fdotnet~2Fcorefx~2Fmaster~2F/test~2Ffunctional~2Fcli~2F/e7308d8b0685078f754361a2b7c6ad2cabaf424a/workItem/System.IO.Compression.Tests) but this has nothing to do with this PR. What shall / can I do here?