@benaadams 2017-12-28
> What shall / can I do here? 

Record and retest

`Windows.7.Amd64.Open-Release-x86`
```
System.IO.Compression.DeflateStreamUnitTests/Dispose_WithUnfinishedWriteAsync

Assert.Throws() Failure
Expected: typeof(System.AggregateException)
Actual:   (No exception was thrown)

   at System.IO.Compression.CompressionStreamUnitTestBase.<Dispose_WithUnfinishedWriteAsync>d__5.MoveNext() in D:\j\workspace\windows-TGrou---f8ac6754\src\Common\tests\System\IO\Compression\CompressionStreamUnitTestBase.cs:line 208
--- End of stack trace from previous location where exception was thrown ---
--- End of stack trace from previous location where exception was thrown ---
--- End of stack trace from previous location where exception was thrown ---
```

@dotnet-bot test Windows x86 Release Build