@benaadams 2017-12-28
> I did some trials for Push too, but the results weren't satisfying. 

In the linked code the test `if ((uint)size >= (uint)array.Length)` doesn't elide the bounds check on the array as the array itself can change (as it returns to the access code after resizing)