@gfoidl 2017-12-28
On the dasm I saw that RCE is not applied, for that reason.
I don't know any way how to make this happen. Except the JIT could trace the new array size (the one from `Array.Resize`, but I'm afraid this kind of flow analysis is going to far.