@benaadams 2017-12-28
Move the Resize path completely out of flow?
```
if
   // ResizeAndPush(val)
else
   // rest of function
```