@gfoidl 2017-12-28
Ah good idea -- one just has to think about it, and that's not always that easy :wink:
I'll try it.