@benaadams 2017-12-28
Issue raised for Deflate fail https://github.com/dotnet/corefx/issues/26089