@mikedn 2018-01-12
The JIT bug that results in incorrect range check elimination has been fixed in https://github.com/dotnet/coreclr/pull/15756. Unfortunately the fix makes range check elimination more conservative when fields are involved so you'll need to add back local variables for `_array`.