@mikedn 2018-01-12
@stephentoub I haven't looked too closely at the jit-diff. I'll check, fortunately it's small enough to look through it.