@mikedn 2018-01-13
> did you happen to notice if there were any existing places in coreclr/corefx that regressed as a result and that we should also fix?

It looks like there are very few cases where we now get new range checks as a result of the JIT fix. I haven't looked at all corefx (the diffs is small but there are so many small and similar changes that are not related to range check elimination that after a while it gets boring) but the very few range checks I have found do not seem very interesting:
* `BitArray`'s constructor has a loop involving an array field. `BitArray` has many other CQ issues so it's not worth going only after that little loop.
* `RegexMatch.Reset` has a similar loop. I'm not familiar with that code but I doubt that its perf is relevant.
* `BigInteger.Equals` has some extra range checks. That code is not in a loop so I doubt that it makes a difference.

The common regression is not related to range checks but to redundant load elimination via CSE. Here's an example from one of `Stack<T>.Push` method:
```diff
--- D:\Projects\diffs\dasmset_13\diff\System.Collections.dasm	2018-01-13 09:19:01.000000000 +-0200
+++ D:\Projects\diffs\dasmset_13\base\System.Collections.dasm	2018-01-13 09:19:00.000000000 +-0200
@@ -50779,12 +50768,12 @@
        mov      ecx, dword ptr [rsi+24]
        mov      rax, gword ptr [rsi+8]
-       mov      eax, dword ptr [rax+8] ; load array length in a register
-       cmp      ecx, eax
+       cmp      ecx, dword ptr [rax+8] ; array length
        jne      SHORT G_M62838_IG05
        lea      rcx, bword ptr [rsi+8]
-       test     eax, eax
+       cmp      dword ptr [rax+8], 0 ; array length again, previously the value in eax was reused
        je       SHORT G_M62838_IG03
        mov      rbx, rcx
-       mov      ebp, eax
+       mov      ecx, dword ptr [rax+8] ; array length again, previously the value in eax was reused
+       mov      ebp, ecx
        shl      ebp, 1
        jmp      SHORT G_M62838_IG04
```
Do such regressions matter? I don't know, I haven't attempted to benchmark this code. It may put up a show in a micro-benchmark but I doubt that it will measurable impact the performance of a real world application, unless that application happens to spend 99% of the time in code affected by this regression :smile: