@gfoidl 2018-01-13
Good (bug fixed) and bad (conservative range checks) 😉 
I've updated the PR.