@karelz 2018-01-27
@stephentoub @mikedn are you ok with merging the change?