@gfoidl 2018-01-30
`Windows.7.Amd64.Open-Debug-x64`

```
Unhandled Exception of Type Xunit.Sdk.EqualException

Assert.Equal() Failure
Expected: NotSignatureValid
Actual:   NotSignatureValid, UntrustedRoot

   at System.Security.Cryptography.X509Certificates.Tests.ChainTests.InvalidSelfSignedSignature() in D:\j\workspace\windows-TGrou---74aa877a\src\System.Security.Cryptography.X509Certificates\tests\ChainTests.cs:line 673
```

@dotnet-bot test Windows x64 Debug Build